Questions Summary of Moodle Grade Report

Calculate questions summary and item analysis of Moodle Grade Report (See Details).

Usage

summary_questions(data, cor_method = c("pearson", "kendall", "spearman"))

Arguments

data

(GradesReport) A data frame of class "GradesReport"

cor_method

(Character) A character string indicating which correlation coefficient is to be used for calculating Discrimination_Index. Default is Pearson's correlation ("pearson"). Other types can be specified such as "kendall", or "spearman", can be abbreviated.

Value

a data frame of overall summary per questions

Details

summary_questions() calculate overall summary grouped by each questions of Moodle Grade Report, and return the result as a data frame. The data frame has a column "Questions" for questions number in the quiz. Other columns are the followings.

Basic summary statistic

n: number of students who answered each questions.

min: student's minimum score of each questions.

max: student's maximum score of each questions.

max_setting: maximum score possible of each questions (from the quiz setting).

mean: mean score of each questions.

SD: standard deviation of each questions score.

Item analysis

Difficulty_Index: Item Difficulty Index (p) is a measure of the proportion of examinees who answered the item correctly. It ranges between 0.0 and 1.0, higher value indicate lower question difficulty, and vice versa.

Discrimination_Index: Item Discrimination Index (r) is a measure of how well an item is able to distinguish between examinees who are knowledgeable and those who are not. It is a pairwise point-biserial correlation between the score of each questions ("Q" columns) and total score of the quiz ("Grade" column). It range between -1.0 to 1.0. Negative values suggest a problem, indicating that score of the particular question is negatively correlated with total quiz score; therefore, revision of the question is suggested.

p.value: A level of significant (p-value) of Discrimination_Index.

p.signif: A Symbol indicating level of significant of Discrimination_Index.

References

How to interpret an item analysis

See also

Examples

# Prepare
grades_df_preped <- prep_grades_report(grades_df)

# Question Summary
summary_questions(grades_df_preped)
#> # A tibble: 9 × 11
#>   Questions     n   min   max max_setting  mean    SD Difficulty_Index
#>   <chr>     <int> <dbl> <dbl>       <dbl> <dbl> <dbl>            <dbl>
#> 1 Q1          271     0     1           1 0.760 0.428            0.760
#> 2 Q2          269     0     1           1 0.941 0.237            0.941
#> 3 Q3          268     0     1           1 0.511 0.501            0.511
#> 4 Q4          269     0     1           1 0.822 0.384            0.822
#> 5 Q5          269     0     1           1 0.743 0.438            0.743
#> 6 Q6          269     0     1           1 0.684 0.466            0.684
#> 7 Q7          269     0     1           1 0.870 0.337            0.870
#> 8 Q8          267     0     1           1 0.708 0.456            0.708
#> 9 Q9          270     0     1           1 0.737 0.441            0.737
#> # … with 3 more variables: Discrimination_Index <dbl>, p.value <dbl>,
#> #   p.signif <chr>