Quiz Summary of Moodle Grade Report

Calculate quiz summary statistics of Moodle Grade Report (See Details).

Usage

summary_quiz(data)

Arguments

data

(GradesReport) A data frame of class "GradesReport" (May expand this to other class in the future)

Value

a one-row data frame of quiz summary statistics

Details

summary_quiz() is a generic function that calculate quiz summary statistics of "GradesReport" data frame.

The result is one row data frame with the following columns:

n: number of students in the quiz

min: student's minimum score

max: student's maximum score

max_quiz: maximum score possible (from the quiz setting)

median: student's median score

Q1: 1st quartile of student's score (see stats::quantile())

Q3: 3rd quartile of student's score (see stats::quantile())

IQR: Interquartile range of student's score (see stats::IQR())

MAD: Median Absolute Deviation of student's score (see stats::mad())

mean: student's mean score

SD: standard deviation of student's score

skewness: a measure of asymmetry, positive values indicate that the tail is on the right, whereas negative values indicate that the tail is on the left (see moments::skewness()).

kurtosis: excess kurtosis, positive values (leptokurtic) indicate that score distribution has a fatter tails (more outliers), whereas negative values (platykurtic) indicate that score distribution has a thinner tails (less outliers). Please note that excess kurtosis = kurtosis - 3. (see moments::skewness() for Pearson's measure of kurtosis).

Cronbach_Alpha: a measure of internal consistency reliability of quiz. It is an average correlation of all question's score ("Q" column). Higher value indicates good overall correlation, and vice versa. (see DescTools::CronbachAlpha())

See also

Examples

# Prepare Quiz
grades_df_preped <- prep_grades_report(grades_df)

# Quiz Summary Statistics
summary_quiz(grades_df_preped)
#> # A tibble: 1 × 14
#>       n   min   max max_quiz median    Q1    Q3   IQR   MAD  mean    SD skewness
#>   <int> <dbl> <dbl>    <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>
#> 1   272     0     9        9      7     5     9     4  2.97  6.70  2.22   -0.884
#> # … with 2 more variables: kurtosis <dbl>, Cronbach_Alpha <dbl>